+++ app/src/androidTest/java/com/tari/android/wallet/FFITestUtil.kt
@@ -45,9 +45,9 @@ class FFITestUtil {
     companion object {
         const val WALLET_DB_NAME: String = "tari_test_db"
         // Matching pair of keys.
-        const val PUBLIC_KEY_HEX_STRING = "28A88213A95BEA688D61293A6DF4EF8B6DA068E95FE515162DA2195AB4E7507C3A"
-        const val PRIVATE_KEY_HEX_STRING = "6259C39F75E27140A652A5EE8AEFB3CF6C1686EF21D27793338D899380E8C801"
-        const val PUBLIC_KEY_EMOJI_ID = "\uD83C\uDF5E\uD83D\uDC7D\uD83D\uDC2C\uD83C\uDF40\uD83D\uDC7E\uD83C\uDFBA\uD83D\uDE3B\uD83C\uDFE6\uD83D\uDC38\uD83C\uDFC0\uD83C\uDF5F\uD83C\uDF79\uD83D\uDC0A\uD83D\uDE97\uD83D\uDE8C\uD83D\uDC36\uD83D\uDC0A\uD83D\uDC5E\uD83C\uDFE6\uD83D\uDE39\uD83C\uDFBE\uD83D\uDE0D\uD83C\uDF44\uD83C\uDF45\uD83C\uDF69\uD83D\uDC60\uD83C\uDF48\uD83C\uDFB9\uD83D\uDC90\uD83D\uDE31\uD83C\uDFAC\uD83D\uDC1E\uD83C\uDF79"
+        const val WALLET_ADDRESS_HEX_STRING = "28A88213A95BEA688D61293A6DF4EF8B6DA068E95FE515162DA2195AB4E7507C3A"
+        const val PRIVATE_KEY_HEX_STRING = "8C7438256641F3A06721AA0DED484AE6EEF48F1C8C95DAF9D6E9D45FA2224D08"
+        const val WALLET_EMOJI_ID = "\uD83C\uDF5E\uD83D\uDC7D\uD83D\uDC2C\uD83C\uDF40\uD83D\uDC7E\uD83C\uDFBA\uD83D\uDE3B\uD83C\uDFE6\uD83D\uDC38\uD83C\uDFC0\uD83C\uDF5F\uD83C\uDF79\uD83D\uDC0A\uD83D\uDE97\uD83D\uDE8C\uD83D\uDC36\uD83D\uDC0A\uD83D\uDC5E\uD83C\uDFE6\uD83D\uDE39\uD83C\uDFBE\uD83D\uDE0D\uD83C\uDF44\uD83C\uDF45\uD83C\uDF69\uD83D\uDC60\uD83C\uDF48\uD83C\uDFB9\uD83D\uDC90\uD83D\uDE31\uD83C\uDFAC\uD83D\uDC1E\uD83C\uDF79"
         val address = NetAddressString("127.0.0.1",80)
 
         fun generateRandomAlphanumericString(len: Int): String {
