+++ app/src/androidTest/java/com/tari/android/wallet/HexStringTests.kt
@@ -9,8 +9,8 @@ class HexStringTests {
     @Test
     fun toString_assertThatGivenConstructorArgumentWasReturned() {
         assertEquals(
-            FFITestUtil.PUBLIC_KEY_HEX_STRING,
-            HexString(FFITestUtil.PUBLIC_KEY_HEX_STRING).toString()
+            FFITestUtil.WALLET_ADDRESS_HEX_STRING,
+            HexString(FFITestUtil.WALLET_ADDRESS_HEX_STRING).toString()
         )
     }
 }
